Refactor crate structure
