+++ Cargo.toml
@@ -1,9 +1,8 @@
 [workspace]
 members = [
-    "analyzer",
     "cache",
-    "gateway",
-    "handler",
+    "event",
+    "interaction",
     "model",
     "raidprotect",
     "util",
