--- analyzer/src/lib.rs
@@ -1,8 +1,0 @@
-//! # RaidProtect message analyzer
-//!
-//! This crate contains the message parser and analyzer used to detect user spam
-//! or filter messages.
-
-mod parser;
-
-pub use parser::{parse_message, ALLOWED_MESSAGES_TYPES};
