+++ event/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "raidprotect-analyzer"
+name = "raidprotect-event"
 version = "0.1.0"
 edition = "2021"
 license = "AGPL-3.0-or-later"
@@ -9,9 +9,11 @@ publish = false
 
 [dependencies]
 raidprotect-cache = { path = "../cache" }
+raidprotect-model = { path = "../model" }
+
+twilight-model = "0.10"
 
 any_ascii = "0.3"
 linkify = "0.8"
-twilight-model = "0.10"
 unicode-segmentation = "1.9"
 url = "2.2"
