+++ event/src/lib.rs
@@ -1,0 +1,7 @@
+//! # RaidProtect events
+//!
+//! This crate contains types used to to handle and respond to Discord events
+//! that the bot receive. Interactions events are not handled in this crate,
+//! as well as events only used to update the cache.
+
+pub mod message;
