--- gateway/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "raidprotect-gateway"
-version = "0.1.0"
-edition = "2021"
-license = "AGPL-3.0-or-later"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-raidprotect-cache = { path = "../cache" }
-raidprotect-handler = { path = "../handler" }
-raidprotect-model = { path = "../model" }
-raidprotect-util = { path = "../util" }
-
-dashmap = "5.1"
-futures = "0.3"
-serde = { version = "1", features = ["derive"] }
-thiserror = "1"
-tokio = { version = "1.17", features = ["macros", "rt-multi-thread", "sync"] }
-tracing = { version = "0.1" }
-
-twilight-model = "0.10"
-twilight-gateway = "0.10"
-twilight-http = "0.10"
