+++ interaction/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "raidprotect-handler"
+name = "raidprotect-interaction"
 version = "0.1.0"
 edition = "2021"
 license = "AGPL-3.0-or-later"
@@ -9,7 +9,7 @@ publish = false
 
 [dependencies]
 raidprotect-model = { path = "../model" }
-raidprotect-analyzer = { path = "../analyzer" }
+raidprotect-util = { path = "../util" }
 
 nanoid = "0.4"
 rosetta-i18n = "0.1"
