+++ interaction/src/embed/error.rs
