+++ interaction/src/embed/mod.rs
