+++ raidprotect/Cargo.toml
@@ -8,12 +8,21 @@ publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-raidprotect-gateway = { path = "../gateway" }
+raidprotect-cache = { path = "../cache" }
+raidprotect-event = { path = "../event" }
+raidprotect-interaction = { path = "../interaction" }
+raidprotect-model = { path = "../model" }
 raidprotect-util = { path = "../util" }
 
 anyhow = "1"
 dotenv = "0.15"
 envy = "0.4"
+futures = "0.3"
 serde = { version = "1", features = ["derive"] }
+thiserror = "1"
 tokio = { version = "1.17", features = ["macros", "rt-multi-thread"] }
 tracing = { version = "0.1" }
+
+twilight-model = "0.10"
+twilight-gateway = "0.10"
+twilight-http = "0.10"
