+++ util/Cargo.toml
@@ -10,6 +10,8 @@ publish = false
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 tokio = { version = "1", features = ["macros", "signal", "sync", "time"] }
+twilight-model = "0.10"
+
 tracing = { version = "0.1", features = ["release_max_level_info"] }
 tracing-appender = "0.2"
 tracing-subscriber = { version = "0.3", features = ["std", "fmt", "ansi"], default-features = false }
