+++ util/src/resource.rs
